// packages/vsce/src/doc/resources/IBundle.ts
/**
 * This program and the accompanying materials are made available under the terms of the
 * Eclipse Public License v2.0 which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-v20.html
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Copyright Contributors to the Zowe Project.
 *
 */

import { IResource } from "./IResource";

export interface IBundle extends IResource {
  bundledir: string;
  bundleid: string;
  enablestatus: string;
  name: string;
  partcount: string;
}